move CustomPlayerListFeature, change renderPlayerNames signature and remove check for empty player list
